feat(core): initial commit
Port scanner to list ports blocked by my ISP

It opens a port on the local machine, give it a timeout and tries to
connect through tor to that said port. If it don't connect in time it
will fail and the port will be identified as closed.

Remember that ports below 1024 are only accessible to Superusers

Still need to implement my version of a socks4 client to connect
directly to the TOR server running on the machine (will be out soon)